Containerisation and temporary integration test
Added Docker containerisation to the maven build bound to the
pre-integration-test goal. Started the Docker image bound to the same
goal, and closed it bound post-integration-test. A healthcheck waited
until the application is available before running integration tests. The
initial integration test is a port of the integration test from the
phase 1 project with modifications to get it to run in junit 5 versus
the HATEOS-compatible Spring Boot webservice.